fix/이미지 관련 로직 image 도메인으로 이동
